Refactor Brs_026 and Brs_028
